+++ docs/examples/react/manner-temp/preview.tsx
@@ -3,13 +3,17 @@ import { MannerTemp } from "seed-design/ui/manner-temp";
 
 export default function MannerTempPreview() {
   return (
-    <VStack gap="x1">
+    <VStack gap="x1" align="flex-end">
       <MannerTemp temperature={12.5} />
-      <MannerTemp temperature={36.2} />
+      <MannerTemp temperature={30} />
+      <MannerTemp temperature={36} />
       <MannerTemp temperature={36.5} />
-      <MannerTemp temperature={41.9} />
-      <MannerTemp temperature={51.9} />
-      <MannerTemp temperature={52} />
+      <MannerTemp temperature={37} />
+      <MannerTemp temperature={40} />
+      <MannerTemp temperature={45} />
+      <MannerTemp temperature={55} />
+      <MannerTemp temperature={65} />
+      <MannerTemp temperature={80} />
     </VStack>
   );
 }