+++ docs/registry/registry-ui.ts
@@ -8,11 +8,11 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "app-screen.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
         {
           path: "app-bar.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -21,7 +21,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "error-state.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -39,7 +39,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "manner-temp.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -48,7 +48,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "manner-temp-badge.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -57,7 +57,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "alert-dialog.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -66,7 +66,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "bottom-sheet.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -75,7 +75,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "action-sheet.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
       deprecated: true,
@@ -85,7 +85,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "extended-action-sheet.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
       deprecated: true,
@@ -95,7 +95,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "avatar.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -104,7 +104,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "pull-to-refresh.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -113,7 +113,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "loading-indicator.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -122,7 +122,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "action-button.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -131,7 +131,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "toggle-button.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -140,7 +140,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "reaction-button.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -149,7 +149,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "callout.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -158,7 +158,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "control-chip.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
       deprecated: true,
@@ -168,7 +168,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "chip.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -177,7 +177,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "checkbox.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -186,7 +186,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "identity-placeholder.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -195,7 +195,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "inline-banner.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
       deprecated: true,
@@ -205,7 +205,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "menu-sheet.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -214,7 +214,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "snackbar.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -223,7 +223,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "help-bubble.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -232,7 +232,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "tabs.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -241,7 +241,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "chip-tabs.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -250,7 +250,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "page-banner.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -259,7 +259,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "progress-circle.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -268,7 +268,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "radio-group.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -277,7 +277,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "select-box.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -286,7 +286,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "segmented-control.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -295,7 +295,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "switch.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -304,7 +304,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "text-field.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -313,7 +313,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "contextual-floating-button.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -322,7 +322,7 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "floating-action-button.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
         },
       ],
     },
@@ -331,7 +331,11 @@ export const registryUI: Registry = {
       snippets: [
         {
           path: "list.tsx",
-          dependencies: { "@seed-design/react": "~0.2.4", "@seed-design/css": "~0.2.4" },
+          dependencies: { "@seed-design/react": "~1.0.0", "@seed-design/css": "~1.0.0" },
+        },
+        {
+          path: "list-header.tsx",
+          dependencies: { "@seed-design/react": "~0.3.0", "@seed-design/css": "~0.3.0" },
         },
       ],
     },