+++ packages/css/recipes/manner-temp.d.ts
@@ -2,7 +2,7 @@ declare interface MannerTempVariant {
   /**
   * @default "l1"
   */
-  level: "l1" | "l2" | "l3" | "l4" | "l5" | "l6";
+  level: "l1" | "l2" | "l3" | "l4" | "l5" | "l6" | "l7" | "l8" | "l9" | "l10";
 }
 
 declare type MannerTempVariantMap = {