+++ packages/css/vars/component/help-bubble.d.ts
@@ -2,11 +2,10 @@ export declare const vars: {
   "base": {
     "enabled": {
       "root": {
-        "color": "var(--seed-color-bg-neutral-solid)",
+        "color": "var(--seed-color-bg-neutral-inverted)",
         "cornerRadius": "var(--seed-radius-r3)",
         "paddingX": "var(--seed-dimension-x3)",
         "paddingY": "var(--seed-dimension-x2_5)",
-        "shadow": "0px 8px 24px 0px #00000029",
         "enterScale": "0.9",
         "enterOpacity": "0",
         "enterDuration": "var(--seed-duration-d4)",
@@ -17,7 +16,7 @@ export declare const vars: {
         "exitTimingFunction": "var(--seed-timing-function-easing)"
       },
       "arrow": {
-        "color": "var(--seed-color-bg-neutral-solid)",
+        "color": "var(--seed-color-bg-neutral-inverted)",
         "width": "12px",
         "height": "8px",
         "cornerRadius": "2px",
@@ -25,19 +24,19 @@ export declare const vars: {
         "padding": "14px"
       },
       "title": {
-        "color": "var(--seed-color-palette-static-white)",
+        "color": "var(--seed-color-fg-neutral-inverted)",
         "fontSize": "var(--seed-font-size-t3)",
         "fontWeight": "var(--seed-font-weight-bold)",
         "lineHeight": "var(--seed-line-height-t3)"
       },
       "description": {
-        "color": "var(--seed-color-palette-static-white)",
+        "color": "var(--seed-color-fg-neutral-inverted)",
         "fontSize": "var(--seed-font-size-t3)",
         "fontWeight": "var(--seed-font-weight-regular)",
         "lineHeight": "var(--seed-line-height-t3)"
       },
       "closeButton": {
-        "color": "var(--seed-color-palette-static-white)",
+        "color": "var(--seed-color-fg-neutral-inverted)",
         "fromRight": "12px",
         "fromTop": "12px",
         "size": "14px",