+++ packages/css/vars/component/switch-mark.d.ts
@@ -1,0 +1,86 @@
+export declare const vars: {
+  "base": {
+    "enabled": {
+      "root": {
+        "color": "var(--seed-color-palette-gray-600)",
+        "cornerRadius": "var(--seed-radius-full)"
+      },
+      "thumb": {
+        "cornerRadius": "var(--seed-radius-full)"
+      }
+    },
+    "disabled": {
+      "root": {
+        "opacity": "0.38"
+      }
+    }
+  },
+  "toneBrand": {
+    "enabled": {
+      "thumb": {
+        "color": "var(--seed-color-palette-static-white)"
+      }
+    },
+    "enabledSelected": {
+      "root": {
+        "color": "var(--seed-color-bg-brand-solid)"
+      }
+    }
+  },
+  "toneNeutral": {
+    "enabled": {
+      "thumb": {
+        "color": "var(--seed-color-fg-neutral-inverted)"
+      }
+    },
+    "enabledSelected": {
+      "root": {
+        "color": "var(--seed-color-bg-neutral-inverted)"
+      }
+    }
+  },
+  "size32": {
+    "enabled": {
+      "root": {
+        "height": "32px",
+        "width": "52px",
+        "paddingX": "3px",
+        "paddingY": "3px"
+      },
+      "thumb": {
+        "height": "26px",
+        "width": "26px",
+        "shadow": "0px 3px 8px 0px #00000026, 0px 1px 3px 0px #0000000f"
+      }
+    }
+  },
+  "size24": {
+    "enabled": {
+      "root": {
+        "height": "24px",
+        "width": "38px",
+        "paddingX": "2px",
+        "paddingY": "2px"
+      },
+      "thumb": {
+        "height": "20px",
+        "width": "20px",
+        "shadow": "0px 3px 8px 0px #00000026, 0px 1px 3px 0px #0000000f"
+      }
+    }
+  },
+  "size16": {
+    "enabled": {
+      "root": {
+        "height": "16px",
+        "width": "26px",
+        "paddingX": "2px",
+        "paddingY": "2px"
+      },
+      "thumb": {
+        "height": "12px",
+        "width": "12px"
+      }
+    }
+  }
+}