+++ packages/qvism-preset/src/vars/color/manner-temp/l7.d.ts
@@ -1,0 +1,2 @@
+export declare const text = "var(--seed-color-manner-temp-l7-text)";
+export declare const bg = "var(--seed-color-manner-temp-l7-bg)";